Reorganized editor components
