+++ src/editor/editors/KeyValueEditor/index.ts
@@ -1,0 +1,1 @@
+export { default as KeyValueEditor } from './KeyValueEditor'
